PR feedback

+++ programs/steward/src/constants.rs
@@ -12,6 +12,6 @@ pub const NUM_EPOCHS_BETWEEN_SCORING_MAX: u64 = 100;
 // Cannot score validators in under 100 slots, to submit 1 instruction per validator
 pub const COMPUTE_SCORE_SLOT_RANGE_MIN: u64 = 100;
 #[cfg(feature = "mainnet-beta")]
-pub const VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH: usize = 520;
+pub const VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH: u64 = 520;
 #[cfg(not(feature = "mainnet-beta"))]
-pub const VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH: usize = 0;
+pub const VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH: u64 = 0;
